pyproject.toml
[tool.poetry]
name = "py-ast-rule-engine"
version = "0.1.0"
description = ""
authors = ["decorator-factory <42166884+decorator-factory@users.noreply.github.com>"]
license = "MPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
